// src/app/modules/vacancy/constants/programming-level.enum.ts
export enum ProgrammingLevelEnum {
  intern = "Практикант",
  junior = "Младший",
  middle = "Средний",
  senior = "Старший",
  lead = "Лид"
}

export enum LevelEnum {
  INTERN = "intern",
  JUNIOR = "junior",
  MIDDLE = "middle",
  SENIOR = "senior",
  LEAD = "lead",
}